fix: accept canonical exponent-form integers on decode across Witness-written artifacts
Witness canonical JSON (canonjson) serializes every number in normalized
scientific form (2578 -> 2.578e3), but strict decoders required plain
integer lexemes, so Witness-written manifests failed Witness's own
downstream readers (pass begin: 'cannot unmarshal number 2.578e3 ...
size of type int64').

Decode-side-only fix: new strictjson.Int64/strictjson.Int accept
mathematically integral JSON numbers including exponent form, rejecting
non-integral and out-of-range values. Applied to freeze.FileEntry.Size,
planning.PlanDocument.BatchSizeMaximum, harness.InventoryFile.Size, and
decode surrogates for adjudication summaries and canonical ledger event
payloads. Canonical serialization and digests are unchanged (round-trip
test pins the 2.578e3 byte form). External relay/user-authored contract
fields deliberately stay strict.

+++ internal/pass/state_validation.go
@@ -512,7 +512,7 @@ func validateFreezeManifestOutput(config Config, manifest freeze.Manifest) (stri
 		if err != nil {
 			return "", diag.Wrap(err, CodeStateInvalid, "snapshot manifest blob could not be read.", diag.WithDetail("path", blobPath))
 		}
-		if int64(len(blobBytes)) != entry.Size {
+		if int64(len(blobBytes)) != int64(entry.Size) {
 			return "", diag.New(CodeStateInvalid, "snapshot manifest blob size does not match file entry.", diag.WithDetail("path", entry.Path), diag.WithDetail("actual_size", len(blobBytes)), diag.WithDetail("expected_size", entry.Size))
 		}
 		if blobDigest := digest.RawBytes(blobBytes); blobDigest != entry.Digest {
